Fix Watch view cannot be used in case of variable added before debugging
Description: When variables are added in the Watch view before debugging, 
all sessions of debugging will send command to Server to get variable value.
But in the session that does not have a source of file including variable, the error 
response from Server is returned. But in amalgamator does not receive this event
in evaluateRequest() function. That makes Watch view hangs for waiting response 
from amalgamator, and Watch view cannot be used then.

Solution: Add exception try...catch to catch the error message return from server when the 
command is not handled successfully. This always returns the response to client if error occurs.